Change token line and column types from int to size_t for better range handling

+++ token/token.h
@@ -97,8 +97,8 @@ typedef struct
 {
     heap_guard_t *value; // The value of the token
     token_type_t type;   // The type of the token (e.g., identifier, keyword, operator)
-    int line;            // The line number where the token was found
-    int column;          // The column number where the token was found
+    size_t line;         // The line number where the token was found
+    size_t column;       // The column number where the token was found
 } token_t;
 
 #endif //FLUENT_TOKEN_H
